Move SchmidtKroneckerProduct to linalg.
It makes much more sense there.
